refactor daterange
