+++ src/sentry/static/sentry/app/components/organizations/timeRangeSelector/dateRange/index.jsx
@@ -1,0 +1,1 @@
+export {default} from './dateRange';
